==> 08_feature_engineering_revisited.ipynb <==
# Feature Engineering

## Set Up

# Importing libraries 
import pandas as pd
import numpy as np 
import matplotlib.pyplot as plt
import seaborn as sns 

df = pd.read_csv('Data/feature_engineered_data.csv', index_col=0)

df.head()

list(df.columns)

# Columns to drop 
cols_to_drop = [

    # Target leakage
    'Price',                 # raw target (using Log_Price instead)
    'Price_per_Seat',
    'Price_per_Year_Age',
    'Brand_Avg_Price',
    'Model_Avg_Price',

    # Raw / redundant mileage features
    'Mileage(miles)',
    'Mileage',
    'Expected_Mileage',
    'Mileage_Delta',

    # Weak / noisy features
    'title',
    'title_lower',
    'Doors',
    'Seats',
    'Previous Owners',
    'Owners_per_Year',
    'Door_Category',
    'Engine_per_Seat',

    # Redundant categorical bins (continuous version would lead to greater insight)
    'Age_Band',
    'Engine_Bin',
    'Usage_Level',
]

# Drop 
df = df.drop(columns=cols_to_drop, errors='ignore')




## Mileage Interaction Features 

df['Mileage_Age_Interaction'] = df['Log_Mileage'] * df['Car_Age']

df['Mileage_per_Age'] = df['Log_Mileage'] / (df['Car_Age'] + 1)


## Depreciation Curve Features

df['Age_3_plus'] = np.maximum(0, df['Car_Age'] - 3)

df['Age_5_plus'] = np.maximum(0, df['Car_Age'] - 5)

df['Age_8_plus'] = np.maximum(0, df['Car_Age'] - 8)


## Engine Interaction Features 

df['Engine_Age'] = df['Engine'] * df['Car_Age']

df['Engine_per_Year'] = df['Engine'] / (df['Car_Age'] + 1)

## Mileage Normalisation Features 

AVG_MILES_PER_YEAR = 12000

df['Mileage_Ratio'] = df['Mileage_per_Year'] / AVG_MILES_PER_YEAR

df['Mileage_Above_Normal'] = df['Mileage_per_Year'] - AVG_MILES_PER_YEAR

## Lifecycle Features 

df['Is_New'] = (df['Car_Age'] <= 2).astype(int)

df['Is_Old'] = (df['Car_Age'] >= 10).astype(int)


## Vehicle Lifespan Features 



# Lifecycle stage categories
df['Lifecycle_Stage'] = pd.cut(
    df['Car_Age'],
    bins=[-1, 2, 5, 10, 100],
    labels=['New', 'Early', 'Mid', 'Late']
)

# Binary lifecycle flags (trees love these)
df['Is_New'] = (df['Car_Age'] <= 2).astype(int)

df['Is_Early_Life'] = ((df['Car_Age'] > 2) & (df['Car_Age'] <= 5)).astype(int)

df['Is_Mid_Life'] = ((df['Car_Age'] > 5) & (df['Car_Age'] <= 10)).astype(int)

df['Is_End_of_Life'] = (df['Car_Age'] > 10).astype(int)

# Remaining useful life proxy (assume 15-year economic lifespan)
df['Remaining_Life'] = np.maximum(0, 15 - df['Car_Age'])

# Depreciation intensity
df['Depreciation_Phase'] = df['Car_Age'] / 15


df.head()

## Brand Features

df['Brand'].unique()


df['Brand'] = df['Brand'].str.strip().str.title()

# Fix known variations
df['Brand'] = df['Brand'].replace({
    'Bmw': 'BMW',
    'Mg': 'MG',
    'Ds': 'DS',
})


premium = [
    'BMW',
    'Mercedes',
    'Audi',
    'Lexus',
    'Porsche',
    'Jaguar',
    'Land Rover',
    'Volvo',
    'Infiniti',
    'Maserati'
]

upper_mid = [
    'Mini',
    'Volkswagen',
    'Toyota',
    'Honda',
    'Mazda',
    'Skoda',
    'Seat',
    'Alfa',
    'Jeep',
    'Subaru'
]

mid = [
    'Ford',
    'Peugeot',
    'Renault',
    'Citroen',
    'Hyundai',
    'Kia',
    'Nissan',
    'Suzuki',
    'Fiat',
    'MG',
    'Chevrolet',
    'Mitsubishi'
]

budget = [
    'Dacia',
    'Proton',
    'Daewoo',
    'Daihatsu',
    'Ssangyong',
    'Chrysler',
    'Dodge',
    'Rover',
    'Saab',
    'Smart',
    'Abarth',
    'DS'
]

# Default assignment
df['Brand_Tier'] = 'Budget'

df.loc[df['Brand'].isin(mid), 'Brand_Tier'] = 'Mid'

df.loc[df['Brand'].isin(upper_mid), 'Brand_Tier'] = 'Upper_Mid'

df.loc[df['Brand'].isin(premium), 'Brand_Tier'] = 'Premium'


## numeric brand strength feature

tier_map = {
    'Budget': 0,
    'Mid': 1,
    'Upper_Mid': 2,
    'Premium': 3
}

df['Brand_Strength'] = df['Brand_Tier'].map(tier_map)


## brand depreciation resistance feature

df['Brand_Residual_Score'] = df['Brand_Strength'] * (1 / (df['Car_Age'] + 1))


df['Is_Premium'] = (df['Brand_Tier'] == 'Premium').astype(int)


df[['Brand', 'Brand_Tier', 'Brand_Strength']].head(20)


## Premium durability

df['Age_Adjusted_Mileage'] = df['Log_Mileage'] / df['Brand_Strength'].replace(0,1)


df = df.drop(['Has_Service_History', 'Is_Premium', 'Is_Family_Car'], axis=1)


df.to_csv('Data/feature_engineered_data_revisited.csv')
